rename: GamutPickerComponent → TonePickerComponent
Renames directory, files, class, selector, and story title.
"Gamut Picker" was a misnomer — reads as picking a gamut, not a
picker bounded by gamut. "Tone Picker" better describes the intent.
