+++ src/app/components/tone-picker/tone-picker.component.scss
